requirements.txt
# Core Django Framework
django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
social-auth-app-django==5.2.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.0.0

# AI/LLM Integration
openai==1.82.1

# HTTP and API utilities
requests==2.32.3

# Document Processing & Export
reportlab==4.2.0
python-docx==1.1.2
markdown2==2.4.13

# Data Processing (for analysis features)
pandas==2.2.2

# File type detection and processing
python-magic==0.4.27

# Development and Production utilities
whitenoise==6.8.2

# Environment configuration
python-decouple==3.8

# Caching and Performance
redis==5.0.1
django-redis==5.4.0

# Package management utilities
setuptools>=68.0.0
